+++ src/assets.ts
@@ -58,6 +58,8 @@ export function assetToName(asset: Asset): string | null {
   if (asset == Asset.PYTH) return "PYTH";
   if (asset == Asset.TIA) return "TIA";
   if (asset == Asset.JTO) return "JTO";
+  if (asset == Asset.ONEMBONK) return "ONEMBONK";
+  if (asset == Asset.SEI) return "SEI";
   if (asset == null) return null; // Some things, like clock callbacks, are for all assets and return asset=null
   return "UNDEFINED";
 }
@@ -72,6 +74,8 @@ export function nameToAsset(name: string): Asset {
   if (name == "PYTH") return Asset.PYTH;
   if (name == "TIA") return Asset.TIA;
   if (name == "JTO") return Asset.JTO;
+  if (name == "SEI") return Asset.SEI;
+  if (name == "ONEMBONK") return Asset.ONEMBONK;
   return Asset.UNDEFINED;
 }
 
@@ -89,6 +93,8 @@ export function toProgramAsset(asset: Asset): any {
   if (asset == Asset.PYTH) return { pyth: {} };
   if (asset == Asset.TIA) return { tia: {} };
   if (asset == Asset.JTO) return { jto: {} };
+  if (asset == Asset.ONEMBONK) return { onembonk: {} };
+  if (asset == Asset.SEI) return { sei: {} };
   return { undefined: {} };
 }
 
@@ -120,6 +126,13 @@ export function fromProgramAsset(asset: any): Asset {
   if (objectEquals(asset, { jto: {} })) {
     return Asset.JTO;
   }
+  if (objectEquals(asset, { onembonk: {} })) {
+    return Asset.ONEMBONK;
+  }
+  if (objectEquals(asset, { sei: {} })) {
+    return Asset.SEI;
+  }
+
   return Asset.UNDEFINED;
 }
 
@@ -151,6 +164,12 @@ export function assetToIndex(asset: Asset): number {
     }
     case Asset.JTO: {
       return 8;
+    }
+    case Asset.ONEMBONK: {
+      return 9;
+    }
+    case Asset.SEI: {
+      return 10;
     }
   }
   return 255; // Undefined is 255 onchain
@@ -185,6 +204,21 @@ export function indexToAsset(index: number): Asset {
     case 8: {
       return Asset.JTO;
     }
+    case 9: {
+      return Asset.ONEMBONK;
+    }
+    case 10: {
+      return Asset.SEI;
+    }
   }
   return Asset.UNDEFINED;
+}
+
+export function assetMultiplier(asset: Asset): number {
+  switch (asset) {
+    case Asset.ONEMBONK: {
+      return 1_000_000;
+    }
+  }
+  return 1;
 }