+++ src/risk-utils.ts
@@ -74,13 +74,11 @@ export function calculatePerpMargin(
   let assetIndex = assets.assetToIndex(asset);
   let initial =
     spotPrice *
-    Math.max(
-      convertNativeBNToDecimal(
-        Exchange.pricing.marginParameters[assetIndex].futureMarginInitial,
-        constants.MARGIN_PRECISION
-      ),
-      0.1 // Temporary override, remove
+    convertNativeBNToDecimal(
+      Exchange.pricing.marginParameters[assetIndex].futureMarginInitial,
+      constants.MARGIN_PRECISION
     );
+
   let maintenance =
     spotPrice *
     convertNativeBNToDecimal(
@@ -320,4 +318,36 @@ export function fakeTrade(
     Math.abs(executionInfo.size)
   );
   return account;
+}
+
+export function addFakeCancelToAccount(
+  marginAccount: CrossMarginAccount,
+  order: types.Order
+) {
+  const assetIndex = assets.assetToIndex(order.asset);
+  const bidAskIndex = order.side == types.Side.BID ? 0 : 1;
+
+  const nativeOrderSize = convertDecimalToNativeLotSize(order.size);
+
+  const cancelOpening = Math.min(
+    marginAccount.productLedgers[assetIndex].orderState.openingOrders[
+      bidAskIndex
+    ].toNumber(),
+    nativeOrderSize
+  );
+  const cancelClosing = nativeOrderSize - cancelOpening;
+
+  marginAccount.productLedgers[assetIndex].orderState.openingOrders[
+    bidAskIndex
+  ] = new anchor.BN(
+    marginAccount.productLedgers[assetIndex].orderState.openingOrders[
+      bidAskIndex
+    ].toNumber() - cancelOpening
+  );
+
+  marginAccount.productLedgers[assetIndex].orderState.closingOrders =
+    new anchor.BN(
+      marginAccount.productLedgers[assetIndex].orderState.closingOrders -
+        cancelClosing
+    );
 }